Separate utility functions from tests file
